BlockMaps.jl

BlockMaps — Module

BlockMaps.jl provides a library of block maps (sliding block code) on subshift of finite type (SFT) aiming for easy-to-use and fast implementation.

Main.BlockMaps.BlockMap — Type

BlockMap{N}(start::Matrix{Int}, target::Matrix{Int}, blocks::Dict{NTuple{N, Int}, Int}, memory::Int)

Construct a N-block map from the given start matrix to the target matrix with the given blocks. If matrix contain only 0 and 1, we treat SFT as vertex shift. If not, we treat SFT as edge shift. For edge shift, the index of the edge is ordered from top to bottom and left to right in the matrix (column major).

Main.BlockMaps.identity — Function

identity(A::Matrix{Int})

Return the identity block map of the given matrix.

Main.BlockMaps.shift — Function

shift(A::Matrix{Int})

Return the shift block map of the given matrix. If A contains only 0 and 1, the function returns the block map on the vertex shift. Otherwise, the function returns the block map on the edge shift.

Main.BlockMaps.marker_endomorphism — Function

marker_automorphism(markers::Tuple{Vararg{Array{Union{Int, Char}, 1}}})

Return the compound marker automorphism on the full two shift.

marker_automorphism(markers::Tuple{Vararg{String}})

Return the compound marker automorphism on the full two shift.

Examples

julia> marker_automorphism("1 1 * 2", "2 * 2 1")
BlockMap([2;;], [2;;], Dict{Tuple{Vararg{Int64}}, Int64}((1, 1, 2, 2, 2) => 1, (1, 1, 2, 1, 2) => 2, (1, 2, 2, 2, 1) => 1, (1, 2, 2, 1, 1) => 2, (2, 1, 1, 2, 1) => 1, (2, 1, 1, 1, 1) => 1, (1, 2, 2, 2, 2) => 2, (1, 2, 2, 1, 2) => 2, (2, 1, 1, 2, 2) => 1, (2, 1, 1, 1, 2) => 1…), 2)

Base.inv — Method

inv(block_map::BlockMap)

Return the inverse block map of the given block map.

Main.BlockMaps.extend — Function

extend(bloc_map::BlockMap, pre::Int, post::Int) -> ::BlockMap

Extend the defining blocks of given block map.

This function return a block map which has definition blocks extended by pre in forward and post in backward.

If pre or post is negative, this function runs shorten instead.

Main.BlockMaps.shorten — Function

shorten(block_map::BlockMap)

Return the block map which is same as the given block map but the length of defining blocks are the shortest.

Examples

f = extend(shift([2;;]), 1, 2)
4-BlockMap: SFT([2;;]) → SFT([2;;]) with memory=0:
(1, 1, 1, 1) => 1
(1, 1, 1, 2) => 1
(1, 1, 2, 1) => 1
(1, 1, 2, 2) => 1
        ⋮
(2, 2, 1, 1) => 2
(2, 2, 1, 2) => 2
(2, 2, 2, 1) => 2
(2, 2, 2, 2) => 2

shorten(f)
1-BlockMap: SFT([2;;]) → SFT([2;;]) with memory=-1:
(1,) => 1
(2,) => 2

shorten(block_map::BlockMap, pre::Int, post::Int)

Return the block map which is same as the given block map but the length of defining blocks are shorten specific length in front and back.

pre and post must be less than shortenable(block_map).

If pre or post is negative, this function runs extend instead.

Examples

f = extend(shift([2;;]), 1, 2)
4-BlockMap: SFT([2;;]) → SFT([2;;]) with memory=0:
(1, 1, 1, 1) => 1
(1, 1, 1, 2) => 1
(1, 1, 2, 1) => 1
(1, 1, 2, 2) => 1
        ⋮
(2, 2, 1, 1) => 2
(2, 2, 1, 2) => 2
(2, 2, 2, 1) => 2
(2, 2, 2, 2) => 2

shorten(f, 1, 1)
2-BlockMap: SFT([2;;]) → SFT([2;;]) with memory=-1:
(1, 1) => 1
(1, 2) => 1
(2, 1) => 2
(2, 2) => 2

Main.BlockMaps.ElementaryStrongShiftEquivalence — Type

ElementaryStrongShiftEquivalence

Elementary strong shift equivalence is a pair of matrices.

Main.BlockMaps.ESSE — Type

Alias of ElementaryStrongShiftEquivalence

Main.BlockMaps.StrongShiftEquivalence — Type

StrongShiftEquivalence

Strong shift equivalence is a sequence of elementary strong shift equivalences.

Main.BlockMaps.SSE — Type

Alias of StrongShiftEquivalence

Main.BlockMaps.reduce! — Function

reduce!(sse::SSE)

Remove identity ESSE components wand return it.

Examples

f = extend(shift([2;;]), 1, 0)
2-BlockMap: SFT([2;;]) → SFT([2;;]) with memory=0:
(1, 1) => 1
(1, 2) => 2
(2, 1) => 1
(2, 2) => 2

sse = _kichens_StrongShiftEquivalence(f)[1]
StrongShiftEquivalence with lag 6:
                               ⎡1 0⎤            ⎡1 1 0 0⎤ ⎡1 0 0 0⎤
                               ⎢0 1⎥  ⎡1 1 0 0⎤ ⎢0 0 1 1⎥ ⎢0 1 0 0⎥  ⎡1 1 0 0⎤
          ⎡1⎤        ⎡1 1 0 0⎤ ⎢1 0⎥  ⎢0 0 1 1⎥ ⎢1 1 0 0⎥ ⎢0 0 1 0⎥  ⎢0 0 1 1⎥
    [1 1] ⎣1⎦  ⎡1 1⎤ ⎣0 0 1 1⎦ ⎣0 1⎦  ⎢1 1 0 0⎥ ⎣0 0 1 1⎦ ⎣0 0 0 1⎦  ⎢1 1 0 0⎥
[2]───────────▶︎⎣1 1⎦─────────────────▶︎⎣0 0 1 1⎦─────────────────────▶︎⎣0 0 1 1⎦─⋯

  ⎡1 1 0 0⎤ ⎡1 0 0 0⎤            ⎡1 0⎤
  ⎢0 0 1 1⎥ ⎢0 1 0 0⎥  ⎡1 1 0 0⎤ ⎢0 1⎥
  ⎢1 1 0 0⎥ ⎢0 0 1 0⎥  ⎢0 0 1 1⎥ ⎢1 0⎥ ⎡1 1 0 0⎤        ⎡1⎤
  ⎣0 0 1 1⎦ ⎣0 0 0 1⎦  ⎢1 1 0 0⎥ ⎣0 1⎦ ⎣0 0 1 1⎦  ⎡1 1⎤ ⎣1⎦ [1 1]
⋯─────────────────────▶︎⎣0 0 1 1⎦─────────────────▶︎⎣1 1⎦───────────▶︎[2]

reduce!(sse)
2-element Vector{ElementaryStrongShiftEquivalence}:
 ElementaryStrongShiftEquivalence([1 1 0 0; 0 0 1 1; 1 1 0 0; 0 0 1 1] => [1 0 0 0; 0 1 0 0; 0 0 1 0; 0 0 0 1])
 ElementaryStrongShiftEquivalence([1 1 0 0; 0 0 1 1; 1 1 0 0; 0 0 1 1] => [1 0 0 0; 0 1 0 0; 0 0 1 0; 0 0 0 1])

sse
StrongShiftEquivalence with lag 4:
                               ⎡1 0⎤            ⎡1 0⎤                                
                               ⎢0 1⎥  ⎡1 1 0 0⎤ ⎢0 1⎥                                
          ⎡1⎤        ⎡1 1 0 0⎤ ⎢1 0⎥  ⎢0 0 1 1⎥ ⎢1 0⎥ ⎡1 1 0 0⎤        ⎡1⎤           
    [1 1] ⎣1⎦  ⎡1 1⎤ ⎣0 0 1 1⎦ ⎣0 1⎦  ⎢1 1 0 0⎥ ⎣0 1⎦ ⎣0 0 1 1⎦  ⎡1 1⎤ ⎣1⎦ [1 1]     
[2]───────────▶︎⎣1 1⎦─────────────────▶︎⎣0 0 1 1⎦─────────────────▶︎⎣1 1⎦───────────▶︎[2]

Main.BlockMaps.reducefirst! — Function

reducefirst!(sse::SSE)

Remove the first identity ESSE component and return it.

Main.BlockMaps.reducelast! — Function

reducelast!(sse::SSE)

Remove the last identity ESSE component and return it.

Main.BlockMaps.StrongShiftEquivalence — Method

StrongShiftEquivalence(::BlockMap) -> ::Tuple{SSE{Int}, Int}

Compute the strong shift equivalence over ℤ₊ (SSE-ℤ₊) of the given block map and how many times the δ(A)⁻¹ must be applied to get the dimension representation of the block map.

You can also use SSE(block_map::BlockMap) instead of StrongShiftEquivalence(block_map::BlockMap).

Main.BlockMaps.dimension_representation — Function

dimension_representation(SSE::StrongShiftEquivalence, [shift_inv::Int])

Compute the dimension representation of the given strong shift equivalence.

Main.BlockMaps.orbitsign_number — Function

orbitsign_number(f::BlockMap, n)

Compute the orbit sign number of the given block map.

The orbit sign number is the sign of the block map as permutation of periodic orbits. os(f, n) can be use for the value to be in ℤ/2ℤ.

Main.BlockMaps.os — Function

0-1 version of orbitsign_number

Main.BlockMaps.gyration_number — Function

gyration_number(f::BlockMap, n)

Compute the gyration number of the given block map. gy(f, n) also can be used.

Main.BlockMaps.gy — Function

alias of gyration_number

Main.BlockMaps.sign_gyration_compatibility_condition — Function

sign_gyration_compatibility_condition(f::BlockMap, n)

Compute the sign gyration compatibility condition of the given block map.

Main.BlockMaps.sgcc — Function

alias of sign_gyration_compatibility_condition