cascade-count.jl

count_cascade — Function

count_cascade(markers::Vector, m::Int)

Count $c_n$, the number of (2, n)-cascade, for all n ≤ m with the following recursive formula:

\[ c_n = \frac{1}{2} \left(p_n - \sum_{\substack{n = 2^i\dot k, \\ i \ge 1}} c_k \right).\]

Here, $p_n$ is the number of periodic orbits of period n in $\Sigma_2 \setminus X$, where the subshift X is defined by markers.

Example

julia> count_cascade([(2, 1, "*", 2, 1)], 8)' # `'` is just for transpose
1×8 adjoint(::Vector{Int64}) with eltype Int64:
 0  0  1  0  1  0  4  6

gcd2_cascade — Function

gcd2_cascade(markers::Vector, m::Int)

Count $c_n$, the number of (2, n)-cascade, for all n ≤ m with the following formula:

\[ c_n = \frac{1}{2n} \sum_{{d|n}} \gcd(2, d) \mu(d) f_{\frac{n}{d}}.\]

Here, $f_n$ is the number of points fixed by $n$-th powered shift $\sigma^n$ in $\Sigma_2 \setminus X$, where the subshift X is defined by markers.

Example

julia> gcd2_cascade([(2, 1, "*", 2, 1)], 8)' # `'` is just for transpose
1×8 adjoint(::Vector{Int64}) with eltype Int64:
 0  0  1  0  1  0  4  6

even1_cascade — Function

Count the number of (2, n)-cascade for all n ≤ m with using the number of periodic orbits with even number of 1's