Pruning.jl

HenonMap — Type

HenonMap(a, b)

Create a Hénon map with parameters a and b defined by

\[H_{a, b}(x, y) = (-x^2 + b y + a, x)\]

The map is area-preserving if $|b| = 1$. When $b = -1$, the map is orientation-preserving.

hyperbolic_fixed_points — Function

hyperbolic_fixed_points(hm::HenonMap)

Return the hyperbolic fixed points of the Hénon map hm.

manifolds — Function

manifolds(hm::HenonMap; num_iterations=16, delta=1e-5, tolerance=1e-3, fixed_pt::Point2o)

Return the stable and unstable manifolds of the hyperbolic fixed points of the Hénon map hm. If fixed_pt_id is provided, only compute manifolds for the specified fixed points (1-based index).

intersections — Function

intersections(lines1, lines2) -> Vector{Tuple{Int,Int,Int,Int,Point2o}}

Compute all intersections between polylines in lines1 and lines2. Returns tuples: (poly1idx, seg1idx, poly2idx, seg2idx, intersection_point).

primary_branch — Function

primary_branch(hm::HenonMap, partition::Function; delta=1e-5, tolerance=1e-3, fixed_pt::Point2o)

Return the primary branches of the stable and unstable manifolds of the Hénon map hm at fixed_pt. The primary branch is defined as the branch that remains in the same partition as the fixed point under iteration.

HomoclinicCode — Type

HomoclinicCode(backward::Vector{Int}, forward::Vector{Int})

A type representing the symbolic code of a homoclinic point, with forward and backward components.

Example

julia> HomoclinicCode([1, 0], [1,0,1])
5-element HomoclinicCode:
(0ᵒᵒ)10.101(0ᵒᵒ)

julia> HomoclinicCode("10.101") # string constructor
5-element HomoclinicCode:
(0ᵒᵒ)10.101(0ᵒᵒ)

symbolic_encoding — Function

symbolic_encoding(hm::HenonMap, pts::Vector{Point2o}, partition::Function, psb::Vector{Point2o}, pub::Vector{Point2o}; iter=15) -> HomoclinicCode

Compute symbolic encoding of each point in pts under Hénon map hm using partition function. The length of backward and forward codes are continued until iteration of the point hit the primary (un-)stable branch, psb (pub). partition should map a Point2o to an integer symbol.

primary_pruning_front — Function

    primary_pruning_front(symb_codes::Vector{HomoclinicCode}) -> Vector{NTuple{4, HomoclinicCode}}

Given a list of homoclinic codes, compute the primary pruning front as a list of blocks. Each block is represented as a tuple of four homoclinic codes (a, b, c, d), where:

a is the top-right corner of the block

b is the bottom-right corner of the block

c is the bottom-left corner of the block

d is the top-left corner of the block

forbidden_words — Function

forbidden_words(blocks::Vector{NTuple{4, HomoclinicCode}}) -> Vector{String}

Given a list of blocks representing the primary pruning front, compute the corresponding forbidden words.

Example

julia> block = (HomoclinicCode("1.01"), HomoclinicCode("1001.01"), HomoclinicCode("1001.010001"), HomoclinicCode("1.010001"))
(HomoclinicCode(1.01), HomoclinicCode(1001.01), HomoclinicCode(1001.010001), HomoclinicCode(1.010001))

julia> forbidden_words([block])
1-element Vector{String}:
 "00101000"

markerize — Function

markerize(words::Vector{String}) -> Vector{String}

Given a list of forbidden words, compute the corresponding markers.

Example

julia> markerize(["0010100", "0011100", "0010101"])
2-element Vector{String}:
 "001010X"
 "0011100"